requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
docutils==0.20.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
iniconfig==2.0.0
jaraco.classes==3.3.1
keyring==24.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.2.0
nh3==0.2.15
openai==1.12.0
packaging==23.2
pkginfo==1.9.6
pluggy==1.4.0
prompt-toolkit==3.0.43
pyairtable==2.2.2
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pytest==8.0.1
python-dotenv==1.0.1
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
setuptools==69.1.0
sniffio==1.3.0
tqdm==4.66.2
twine==5.0.0
typing_extensions==4.9.0
urllib3==2.2.1
validators==0.22.0
wcwidth==0.2.13
wheel==0.42.0
zipp==3.17.0
